imp(loading): Show loading state until data is available
