+++ routing-and-rendering-deep-dive/app/(content)/archive/@archive/[[...filters]]/contracts.ts
@@ -1,0 +1,9 @@
+interface IArchivePage {
+  params: Promise<{ filters: string[] }>;
+}
+
+interface IFilteredContent {
+  year: string;
+}
+
+export type { IArchivePage, IFilteredContent };
