+++ routing-and-rendering-deep-dive/app/(content)/archive/loading.tsx
@@ -1,0 +1,3 @@
+export default function Loading() {
+   return <p>Loading archived news...</p>;
+}
