+++ routing-and-rendering-deep-dive/app/(content)/news/loading.tsx
@@ -1,0 +1,3 @@
+export default function Loading() {
+   return <p>Loading news...</p>;
+}
